feat: notification timestamp label + file icon HiDPI fix
- Render a small light-gray HH:MM time label at each window's
  top-right corner via GtkOverlay (manually positioned on configure,
  since GTK3 pins overlay children top-left); pass-through so it never
  steals hover/click.
- Thread the generation timestamp through Pending -> window content ->
  HistoryEntry (kept as `created` for history re-display), and format
  it with glib::DateTime::from_unix_local.
- Fix file icon HiDPI rendering: GtkImage draws a pixbuf at its pixel
  size in logical px then re-scales, so a `target*scale` pixbuf
  double-scaled. Now wrap the scaled pixbuf in a cairo surface carrying
  the device scale so it draws at `target` logical / `target*scale`
  physical, crisp on HiDPI.
- Drop a redundant i64->i32 cast on the cmd0 progress hint.
- Stop tracking the compiled Python integration-test pyc (gitignore).

All 39 unit tests pass.

+++ src/dbus.rs
@@ -350,7 +350,7 @@ impl Cmd0 {
                 );
                 m.insert("timeout".to_string(), Value::from(h.expire_timeout as i64));
                 m.insert("timestamp".to_string(), Value::from(h.timestamp as i64));
-                m.insert("progress".to_string(), Value::from(h.value.unwrap_or(-1) as i32));
+                m.insert("progress".to_string(), Value::from(h.value.unwrap_or(-1)));
                 m
             })
             .collect()
